Moved BackendTester to backend
